+++ runtime/musl-lkl/lkl/tools/perf/bench/sched-pipe.c
@@ -9,7 +9,10 @@
  *  http://people.redhat.com/mingo/cfs-scheduler/tools/pipe-test-1m.c
  * Ported to perf by Hitoshi Mitake <mitake@dcl.info.waseda.ac.jp>
  */
+#include "../perf.h"
+#include "../util/util.h"
 #include <subcmd/parse-options.h>
+#include "../builtin.h"
 #include "bench.h"
 
 #include <unistd.h>
@@ -156,7 +159,7 @@ int bench_sched_pipe(int argc, const char **argv)
 		result_usec += diff.tv_usec;
 
 		printf(" %14s: %lu.%03lu [sec]\n\n", "Total time",
-		       (unsigned long) diff.tv_sec,
+		       diff.tv_sec,
 		       (unsigned long) (diff.tv_usec / USEC_PER_MSEC));
 
 		printf(" %14lf usecs/op\n",
@@ -168,7 +171,7 @@ int bench_sched_pipe(int argc, const char **argv)
 
 	case BENCH_FORMAT_SIMPLE:
 		printf("%lu.%03lu\n",
-		       (unsigned long) diff.tv_sec,
+		       diff.tv_sec,
 		       (unsigned long) (diff.tv_usec / USEC_PER_MSEC));
 		break;
